+++ s2p_lib.h
@@ -9,15 +9,15 @@ extern png_structp png_ptr;
 
 /* In c.reduce: */
 
-char *reduceto8 (long *image, long *mask,
+char *reduceto8 (uint32_t *image, uint32_t *mask,
   rgb_t **ipalette, int *entries, png_color_16 *bkgd);
 
-char *reduceto8_alpha(long *image, char **mask,
+char *reduceto8_alpha(uint32_t *image, char **mask,
   rgb_t **ipalette, int *entries, png_color_16 *bkgd);
 
-int reducetogrey (long *image, char *mask, int rgba);
+int reducetogrey (uint32_t *image, char *mask, int rgba);
 
-png_color_8 significantbits (long *image);
+png_color_8 significantbits (uint32_t *image);
 
 /* In c.freqsort: */
 
@@ -29,6 +29,6 @@ void frequency_sort (char *const image, rgb_t *const palette,
 extern int trim_mask_8 (char *const image, char *const alpha, int maskcolour);
 extern int trim_mask_8a (grey_t *const image);
 extern int trim_mask_24a (rgb_t *const image);
-extern int trim_mask_24 (rgb_t *const image, char *const alpha, long maskcolour);
+extern int trim_mask_24 (rgb_t *const image, char *const alpha, uint32_t maskcolour);
 
 #endif
